Switch to PDAs for fee receiver

+++ programs/castle-lending-aggregator/src/lib.rs
@@ -22,7 +22,7 @@ pub mod castle_lending_aggregator {
         ctx: Context<Initialize>,
         _bumps: InitBumpSeeds,
         strategy_type: StrategyType,
-        fee_bps: u64,
+        fee_bps: u8,
     ) -> ProgramResult {
         instructions::init::handler(ctx, _bumps, strategy_type, fee_bps)
     }
